+++ workflow/scripts/modify_prenetwork.py
@@ -1055,6 +1055,20 @@ def force_connection_nep_offshore(n, current_year):
     # Load shapes and projects
     offshore = pd.read_csv(snakemake.input.offshore_connection_points, index_col=0)
 
+    if int(snakemake.params.offshore_nep_force["delay_years"]) != 0:
+        # Modify 'Inbetriebnahmejahr' by adding the delay years for rows where 'Inbetriebnahmejahr' > 2025
+        offshore.loc[
+            offshore["Inbetriebnahmejahr"] > 2025, "Inbetriebnahmejahr"
+        ] += int(snakemake.params.offshore_nep_force["delay_years"])
+        logger.info(
+            f"Delaying NEP offshore connection points by {snakemake.params.offshore_nep_force['delay_years']} years."
+        )
+        # This is a hack s.t. for CurPol and WorstCase the 2030 projects are delayed to the 2035 period, but the later projects are ignored
+        offshore.loc[offshore["Inbetriebnahmejahr"] > 2031, "Inbetriebnahmejahr"] += 5
+        logger.info(
+            "Delaying NEP offshore connection points after 2031 by another 5 years."
+        )
+
     goffshore = gpd.GeoDataFrame(
         offshore,
         geometry=gpd.points_from_xy(offshore.lon, offshore.lat),